# openapi/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: oojuniin@outlook.com
    name: Junior Lima
  description: "Contrato REST da Workbox API. Este documento é a fonte da verdade\
    \ para qualquer client (frontend, agentes de IA) que consuma esta API — não assuma\
    \ comportamento não descrito aqui."
  title: Workbox API
  version: 0.0.1
servers:
- description: Relativo ao host onde a API estiver publicada
  url: /
security:
- bearer-jwt: []
paths:
  /api/v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserApiLoginCredentialsDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses:
        "200":
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/me:
    get:
      operationId: me
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserApiDTO"
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/password:
    put:
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDTO"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      parameters:
      - in: query
        name: refreshToken
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
          description: OK
      tags:
      - auth-controller
  /api/v1/auth/reset-password:
    post:
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordDTO"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - auth-controller
  /api/v1/role:
    get:
      operationId: findAll
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleDTO"
          description: OK
      tags:
      - role-controller
    post:
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleDTO"
          description: Created
      tags:
      - role-controller
  /api/v1/role/{id}:
    delete:
      operationId: delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      tags:
      - role-controller
    get:
      operationId: findById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleDTO"
          description: OK
      tags:
      - role-controller
    put:
      operationId: update_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleDTO"
          description: OK
      tags:
      - role-controller
  /api/v1/user/find-all:
    get:
      operationId: findAll_2
      parameters:
      - in: query
        name: User API
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollectionModelUserApiDTO"
          description: OK
      tags:
      - user-api-controller
  /api/v1/user/pageable:
    get:
      operationId: findAll_1
      parameters:
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - in: query
        name: User API
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserApiDTO"
          description: OK
      tags:
      - user-api-controller
  /api/v1/user/save:
    post:
      operationId: save
      parameters:
      - in: query
        name: User API
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserApiInsertOrUpdateDTO"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserApiDTO"
          description: Created
      tags:
      - user-api-controller
  /api/v1/user/update:
    put:
      operationId: update
      parameters:
      - in: query
        name: User API
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserApiInsertOrUpdateDTO"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserApiDTO"
          description: OK
      tags:
      - user-api-controller
  /api/v1/user/{id}:
    delete:
      operationId: delete_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
      tags:
      - user-api-controller
    get:
      operationId: findById_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: User API
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EntityModelUserApiDTO"
          description: OK
      tags:
      - user-api-controller
components:
  schemas:
    ChangePasswordDTO:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 100
          minLength: 8
      required:
      - currentPassword
      - newPassword
    CollectionModelUserApiDTO:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            userApiDTOList:
              type: array
              items:
                $ref: "#/components/schemas/UserApiDTO"
        _links:
          $ref: "#/components/schemas/Links"
    EntityModelUserApiDTO:
      type: object
      properties:
        _links:
          $ref: "#/components/schemas/Links"
        email:
          type: string
        enabled:
          type: boolean
        id:
          type: string
          format: uuid
        username:
          type: string
    ForgotPasswordDTO:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
      - email
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    Link:
      type: object
      properties:
        deprecation:
          type:
          - string
          - "null"
        href:
          type: string
        hreflang:
          type:
          - string
          - "null"
        name:
          type:
          - string
          - "null"
        profile:
          type:
          - string
          - "null"
        templated:
          type: boolean
        title:
          type:
          - string
          - "null"
        type:
          type:
          - string
          - "null"
    Links:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/Link"
    PageUserApiDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserApiDTO"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    ResetPasswordDTO:
      type: object
      properties:
        newPassword:
          type: string
          maxLength: 100
          minLength: 8
        token:
          type: string
          minLength: 1
      required:
      - newPassword
      - token
    Role:
      type: object
      properties:
        authority:
          type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        deletedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserApi"
          uniqueItems: true
      required:
      - authority
    RoleDTO:
      type: object
      properties:
        authority:
          type: string
          minLength: 1
        id:
          type: integer
          format: int64
      required:
      - authority
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    UserApi:
      type: object
      properties:
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
        authorities:
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        credentialsNonExpired:
          type: boolean
        deletedAt:
          type: string
          format: date-time
        email:
          type: string
          format: email
        enabled:
          type: boolean
        failedLoginAttempts:
          type: integer
          format: int32
        id:
          type: string
          format: uuid
        isAccountNonExpired:
          type: boolean
        isAccountNonLocked:
          type: boolean
        isCredentialsNonExpired:
          type: boolean
        isEnabled:
          type: boolean
        lockedUntil:
          type: string
          format: date-time
        password:
          type: string
          writeOnly: true
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
          uniqueItems: true
        tokenVersion:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        username:
          type: string
          maxLength: 50
          minLength: 5
      required:
      - username
    UserApiDTO:
      type: object
      properties:
        _links:
          $ref: "#/components/schemas/Links"
        email:
          type: string
        enabled:
          type: boolean
        id:
          type: string
          format: uuid
        username:
          type: string
    UserApiInsertOrUpdateDTO:
      type: object
      properties:
        email:
          type: string
        id:
          type: string
          format: uuid
        isEnabled:
          type: boolean
        password:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
          uniqueItems: true
        username:
          type: string
    UserApiLoginCredentialsDTO:
      type: object
      properties:
        password:
          type: string
        username:
          type: string
  securitySchemes:
    bearer-jwt:
      bearerFormat: JWT
      name: bearer-jwt
      scheme: bearer
      type: http